working

+++ ft_putstr_fd.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_putstr_fd.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yalaatik <yalaatik@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/27 11:04:36 by yalaatik          #+#    #+#             */
+/*   Updated: 2025/01/27 11:38:42 by yalaatik         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "ft_printf.h"
 
 int	ft_putstr_fd(char const*s, int fd)
 {
 	if (!s)
 		return (ft_putstr_fd("(null)", fd));
-	return ((int)write(fd, s, ft_strlen(s)));
+	return (write(fd, s, ft_strlen(s)));
 }
